pyproject.toml
[project]
name = "operation-subjects"
version = "0.1.0"
description = "A TUI-based bioinformatics suite."
authors = [
    { name = "MD. Arshad", email = "arshad10867c@gmail.com" }
]
requires-python = ">=3.8"
dependencies = [
    "rich",
    "biopython",
    "matplotlib",
    "numpy",
    "scipy"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
